+++ tests/Relation/BelongsToTest.php
@@ -13,45 +13,66 @@ use Pulsar\Relation\BelongsTo;
 
 class BelongsToTest extends PHPUnit_Framework_TestCase
 {
+    public static $adapter;
+
     public static function setUpBeforeClass()
     {
-        $adapter = Mockery::mock('Pulsar\Adapter\AdapterInterface');
-
-        $adapter->shouldReceive('queryModels')
-                ->andReturn([['id' => 11]]);
-
-        Model::setAdapter($adapter);
+        self::$adapter = Mockery::mock('Pulsar\Adapter\AdapterInterface');
+        Model::setAdapter(self::$adapter);
     }
 
     public function testInitQuery()
     {
-        $model = new TestModel2();
-        $model->test_model_id = 10;
+        $post = new Post(['category_id' => 10]);
 
-        $relation = new BelongsTo('TestModel', 'id', 'test_model_id', $model);
+        $relation = new BelongsTo($post, 'category_id', 'Category', 'id');
 
-        $this->assertEquals(['id' => 10], $relation->getQuery()->getWhere());
-        $this->assertEquals(1, $relation->getQuery()->getLimit());
+        $query = $relation->getQuery();
+        $this->assertInstanceOf('Category', $query->getModel());
+        $this->assertEquals(['id' => 10], $query->getWhere());
+        $this->assertEquals(1, $query->getLimit());
     }
 
     public function testGetResults()
     {
-        $model = new TestModel2();
-        $model->test_model_id = 10;
+        $post = new Post(['category_id' => 10]);
 
-        $relation = new BelongsTo('TestModel', 'id', 'test_model_id', $model);
+        $relation = new BelongsTo($post, 'category_id', 'Category', 'id');
+
+        self::$adapter->shouldReceive('queryModels')
+                      ->andReturn([['id' => 11]]);
 
         $result = $relation->getResults();
-        $this->assertInstanceOf('TestModel', $result);
+        $this->assertInstanceOf('Category', $result);
         $this->assertEquals(11, $result->id());
     }
 
     public function testEmpty()
     {
-        $model = new TestModel2(['test_model_id' => null]);
+        $post = new Post(['category_id' => null]);
 
-        $relation = new BelongsTo('TestModel', 'id', 'test_model_id', $model);
+        $relation = new BelongsTo($post, 'category_id', 'Category', 'id');
 
         $this->assertNull($relation->getResults());
+    }
+
+    public function testCreate()
+    {
+        $post = new Post(['category_id' => null]);
+
+        $relation = new BelongsTo($post, 'category_id', 'Category', 'id');
+
+        self::$adapter->shouldReceive('createModel')
+                      ->andReturn(true);
+
+        self::$adapter->shouldReceive('getCreatedID')
+                     ->andReturn(1);
+
+        $category = $relation->create(['test' => true]);
+
+        $this->assertInstanceOf('Category', $category);
+        $this->assertEquals(1, $post->category_id);
+        $this->assertEquals(true, $category->test);
+        $this->assertTrue($category->persisted());
     }
 }